refactor: rename 'MoreFragment'
