+++ ICT4DNews/app/src/main/res/layout/fragment_ict4d.xml
@@ -12,7 +12,7 @@
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:paddingBottom="16dp"
-                tools:context=".screens.more.MoreFragment2">
+                tools:context=".screens.more.MoreFragment">
 
             <com.google.android.material.textview.MaterialTextView
                     android:id="@+id/ict4d_text"
