+++ ICT4DNews/app/src/main/res/layout/fragment_more.xml
@@ -9,14 +9,14 @@
 
         <variable
                 name="fragment"
-                type="at.ict4d.ict4dnews.screens.more.MoreFragment" />
+                type="at.ict4d.ict4dnews.screens.more.MoreFragmentOld" />
     </data>
 
     <LinearLayout
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:orientation="vertical"
-            tools:context=".screens.more.MoreFragment">
+            tools:context=".screens.more.MoreFragmentOld">
 
         <include
                 android:id="@+id/include"
