+++ ICT4DNews/app/src/main/res/navigation/nav_graph.xml
@@ -26,7 +26,7 @@
 
     <fragment
             android:id="@+id/moreFragment"
-            android:name="at.ict4d.ict4dnews.screens.more.MoreFragment2">
+            android:name="at.ict4d.ict4dnews.screens.more.MoreFragment">
 
         <action
                 android:id="@+id/action_actionMore_to_settingsFragment"
